# apps/worker/docs/swagger.yaml
...
info:
  contact: {}
  description: LLM API Gateway — aggregate providers, manage load balancing, and track
    usage/costs.
  title: Wheel API
  version: 1.0.0
paths:
...
  /api/v1/model/channel:
    get:
      produces:
        - application/json
      responses:
        "200":
          description: "{success: true, data: []object}"
          schema:
            type: object
      security:
        - BearerAuth: []
      summary: List models grouped by channel
      tags:
        - Models
  /api/v1/model/create:
    post:
      consumes:
        - application/json
      parameters:
        - description: Model name and prices
          in: body
          name: body
          required: true
          schema:
            type: object
      produces:
        - application/json
      responses:
        "200":
          description: "{success: true, data: LLMPrice}"
          schema:
            type: object
        "400":
          description: "{success: false, error: string}"
          schema:
            type: object
        "409":
          description: "{success: false, error: string}"
          schema:
            type: object
      security:
        - BearerAuth: []
      summary: Create a model price entry
      tags:
        - Models
  /api/v1/model/delete:
    post:
      consumes:
        - application/json
      parameters:
        - description: Model ID
          in: body
          name: body
          required: true
          schema:
            type: object
      produces:
        - application/json
      responses:
        "200":
          description: "{success: true}"
          schema:
            type: object
        "400":
          description: "{success: false, error: string}"
          schema:
            type: object
      security:
        - BearerAuth: []
      summary: Delete a model price entry
      tags:
        - Models
  /api/v1/model/last-update-time:
    get:
      produces:
        - application/json
      responses:
        "200":
          description: "{success: true, data: {lastUpdateTime: string}}"
          schema:
            type: object
      security:
        - BearerAuth: []
      summary: Get last price sync timestamp
      tags:
        - Models
  /api/v1/model/list:
    get:
      produces:
        - application/json
      responses:
        "200":
          description: "{success: true, data: {models: []LLMPrice}}"
          schema:
            type: object
      security:
        - BearerAuth: []
      summary: List all model prices
      tags:
        - Models
  /api/v1/model/metadata:
    get:
      produces:
        - application/json
      responses:
        "200":
          description: "{success: true, data: map[string]modelMeta}"
          schema:
            type: object
      security:
        - BearerAuth: []
      summary: Get model metadata (provider, logo, etc.)
      tags:
        - Models
  /api/v1/model/metadata/refresh:
    post:
      produces:
        - application/json
      responses:
        "200":
          description: "{success: true, data: {count: int}}"
          schema:
            type: object
      security:
        - BearerAuth: []
      summary: Refresh model metadata from models.dev
      tags:
        - Models
  /api/v1/model/update:
    post:
      consumes:
        - application/json
      parameters:
        - description: Model ID and optional fields to update
          in: body
          name: body
          required: true
          schema:
            type: object
      produces:
        - application/json
      responses:
        "200":
          description: "{success: true}"
          schema:
            type: object
        "400":
          description: "{success: false, error: string}"
          schema:
            type: object
      security:
        - BearerAuth: []
      summary: Update a model price entry
      tags:
        - Models
  /api/v1/model/update-price:
    post:
      produces:
        - application/json
      responses:
        "200":
          description: "{success: true, data: {synced: int, updated: int}}"
          schema:
            type: object
      security:
        - BearerAuth: []
      summary: Sync model prices from models.dev
      tags:
        - Models
...
  /api/v1/stats/model:
    get:
      produces:
        - application/json
      responses:
        "200":
          description: "{success: true, data: []ModelStatsItem}"
          schema:
            type: object
      security:
        - BearerAuth: []
      summary: Get per-model statistics
      tags:
        - Stats
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"